requirements.txt
pandas
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
seaborn
requests
matplotlib
numpy
folium
google-generativeai
streamlit-folium
folium
openrouteservice
plotly
geopy
langchain
langchain-community
langchain-core
langchain-google-genai
bcrypt
captcha
python-dotenv
tqdm
chromadb
faiss-cpu
pypdf
google-genai
PIL